Succesfully able to input scores
