+++ dartbot_redux/lib/backend/match_engine/match_engine.dart
@@ -3,164 +3,81 @@ import 'package:flutter/foundation.dart';  // for ChangeNotifier
 import 'package:dartbot_redux/backend/match_engine/dart_player.dart';
 import 'package:dartbot_redux/backend/match_engine/match_logic.dart';
 
-class MatchEngine {
+class MatchEngine extends ChangeNotifier{
   DartPlayer player1;
   DartPlayer player2;
   final MatchLogic matchRules;
-  int onThrow = 2;
-  int onThrowSet = 2;
+  int throwing = 1;
+  int onThrow = 1;
+  int onThrowSet = 1;
 
 
   MatchEngine(this.player1, this.player2, this.matchRules);
 
-  void newLeg() {
-        this.player1.score = matchRules.getStartScore();
-        this.player1.stats.dartsThrownLeg = 0;
-        this.player2.score = matchRules.getStartScore();
-        this.player2.stats.dartsThrownLeg = 0;
-        if (this.onThrow == 1) {
-            this.onThrow = 2;
-        } else {
-            this.onThrow = 1;
-        }
-        if (this.player1.legs == 0 && this.player2.legs == 0) {
-            if (this.onThrowSet == 1) {
-                this.onThrow = 2;
-                this.onThrowSet = 2;
-            } else {
-                this.onThrow = 1;
-                this.onThrowSet = 1;
-            }
-        }
-    }
-
-    void playLeg(bool ifQuickSim) {
-        this.newLeg();
-        int playerToThrow = this.onThrow;
-        while (this.player1.score > 0 && this.player2.score > 0) {
-            if (playerToThrow == 1) {
-                bool doubleInOpener = this.matchRules.ifDoubleIn() && this.player1.score == this.matchRules.getStartScore();
-                this.player1.visitThrow(this.matchRules.ifDoubleOut(), doubleInOpener);
-                playerToThrow = 2;
-            } else {
-                bool doubleInOpener = this.matchRules.ifDoubleIn() && this.player2.score == this.matchRules.getStartScore();
-                this.player2.visitThrow(this.matchRules.ifDoubleOut(), doubleInOpener);
-                playerToThrow = 1;
-            }
-
-            if (!ifQuickSim) {
-                if (matchRules.isSetPlay) {
-                    print(this.player1.toStringSetPlay());
-                    print(this.player2.toStringSetPlay());
-                } else {
-                    print(this.player1.toString());
-                    print(this.player2.toString());
-                }
-                print('');
-            }
-        }
-
-        DartPlayer? hasWonLeg = this.ifWonLeg();
-        if (hasWonLeg != null) {
-            if (hasWonLeg.equals(this.player1)) {
-                if (!ifQuickSim) print("${this.player1.name} has won the leg!");
-                this.player1.updateBestWorstLegs();
-            } else {
-                if (!ifQuickSim) print("${this.player2.name} has won the leg!");
-                this.player2.updateBestWorstLegs();
-            }
-            
-            DartPlayer? hasWonSet = this.ifWonSet();
-            if (hasWonSet != null) {
-                if (hasWonSet.equals(this.player1)) {
-                    if (!ifQuickSim) print("${this.player1.name} has won the set!");
-                } else {
-                    if (!ifQuickSim) print("${this.player2.name} has won the set!");
-                }
-            }
-        } 
-
-        if (!ifQuickSim) {
-            if (matchRules.isSetPlay) {
-                print(this.toStringSetPlay());
-            } else {
-                print(this.toString());
-            }
-        }
-    }
-
-  DartPlayer? ifWonSet() {
-        if (!matchRules.isSetPlay) {
-            return null;
-        }
-
-        if (this.player1.legs >= matchRules.getLegLimit()) {
-            this.player1.sets++;
-            this.player1.legs = 0;
-            this.player2.legs = 0;
-            return this.player1;
-        } else if (this.player2.legs >= matchRules.getLegLimit()) {
-            this.player2.sets++;
-            this.player1.legs = 0;
-            this.player2.legs = 0;
-            return this.player2;
-        }
-        return null;
-    }
-
-    DartPlayer? ifWonLeg() {
-        if (this.player1.score <= 0) {
-            this.player1.legs++;
-            this.player2.score = 0;
-            return this.player1;
-        } else if (this.player2.score <= 0) {
-            this.player2.legs++;
-            this.player1.score = 0;
-            return this.player2;
-        }
-
-        return null;
-    }
-
-    DartPlayer? ifWinner(DartPlayer player) {
-        if (matchRules.isSetPlay) {
-            if (player.sets >= matchRules.getSetLimit()) {
-                return player;
-            } 
-            return null;
-        } else {
-            if (player.legs >= matchRules.getLegLimit()) {
-                return player;
-            }
-            return null;
-        }
-    }
-
-    @override
-    String toString() {
-        return "Current Result:\nLeg|Scr\n(${player1.legs}) (${player1.score}) ${player1.name}\n(${player2.legs}) (${player2.score}) ${player2.name}";
-
-        
-    }
-
-     String toStringSetPlay() {
-        return "Current Result:\nSet|Leg|Scr\n(${player1.sets}) (${player1.legs}) (${player1.score}) ${player1.name}\n(${player2.sets}) (${player2.legs}) (${player2.score}) ${player2.name}";
-    }
-
-    @override
-    bool operator ==(Object other) {
-      if (identical(this, other)) return true;
-      if (other is! MatchEngine) return false;
-  
-      return this.player1 == other.player1 &&
-              this.player2 == other.player2 &&
-              this.matchRules == other.matchRules;
-
-    }
-
-    @override
-    int get hashCode {
-      return this.player1.hashCode + this.player2.hashCode;
+  void initMatch() {
+    newLeg();
+    onThrow = 1;
+    onThrowSet = 1;
+    throwing = 1;
   }
+
+  void newLeg() {
+        player1.score = matchRules.getStartScore();
+        player1.stats.dartsThrownLeg = 0;
+        player2.score = matchRules.getStartScore();
+        player2.stats.dartsThrownLeg = 0;
+        if (onThrow == 1) {
+            onThrow = 2;
+        } else {
+            onThrow = 1;
+        }
+        if (player1.legs == 0 && player2.legs == 0) {
+            if (onThrowSet == 1) {
+                onThrow = 2;
+                onThrowSet = 2;
+            } else {
+                onThrow = 1;
+                onThrowSet = 1;
+            }
+        }
+        throwing = onThrow;
+    }
+
+  bool visitThrow(int pointsScored) {
+    DartPlayer playerThrowing;
+    if (throwing == 1) {
+      playerThrowing = player1;
+    } else {
+      playerThrowing = player2;
+    }
+
+    String errorString = "";
+    bool successfulThrow = playerThrowing.visitThrow(pointsScored, 
+                                                     matchRules.doubleOut, 
+                                                     matchRules.doubleIn, 
+                                                     errorString);
     
+    if (!successfulThrow) {
+      /// Is going to be replaced with a pop-up
+      print("ERROR: $errorString");
+      return false;
+    }
+    
+    /// TODO: ADD POP-UP FOR DOUBLES AND DARTS AT CHECKOUT
+    
+    
+    playerThrowing.score -= pointsScored;
+    playerThrowing.dartThrow(pointsScored, matchRules.doubleOut, 0);
+
+    // Hands the "turn" to the next player
+    if (throwing == 1) {
+      throwing = 2;
+    } else {
+      throwing = 1;
+    }
+
+    notifyListeners();  
+    return true;
+  }
+
 }
